Move PLD sim into its own folder
