+++ packages/core/src/sims/tank/pld/pld_usage_sim_no_sks.ts
