+++ packages/frontend/src/scripts/sims/sim_guis.ts
@@ -24,9 +24,9 @@ import { rprSheetSpec } from "@xivgear/core/sims/melee/rpr/rpr_sheet_sim";
 import { samSpec } from "@xivgear/core/sims/melee/sam/sam_lv100_sim";
 import { vprSheetSpec } from "@xivgear/core/sims/melee/vpr/vpr_sheet_sim";
 import { dncDtSheetSpec } from "@xivgear/core/sims/ranged/dnc_sim";
-import { pldUsageSimSpec } from "@xivgear/core/sims/tank/pld_usage_sim_no_sks";
+import { pldUsageSimSpec } from "@xivgear/core/sims/tank/pld/pld_usage_sim_no_sks";
 import { drkSpec } from "@xivgear/core/sims/tank/drk/drk_lv100_sim";
-import { pldSKSSheetSpec } from "@xivgear/core/sims/tank/pldsks_sim";
+import { pldSKSSheetSpec } from "@xivgear/core/sims/tank/pld/pldsks_sim";
 import { BluBreath60Spec } from "@xivgear/core/sims/blu/blu_breath60";
 import { BluFlame120Spec } from "@xivgear/core/sims/blu/blu_flame120";
 import { BluFlame60Spec } from "@xivgear/core/sims/blu/blu_flame60";
